% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/href.R
\name{href_book}
\alias{href_book}
\alias{href_github}
\alias{href_pkgdown}
\title{Get useful openair links}
\usage{
href_book(open = FALSE, ...)

href_github(open = FALSE, ...)

href_pkgdown(package = "openair", open = FALSE, ...)
}
\arguments{
\item{open}{If \code{TRUE} will use \code{\link[utils:browseURL]{utils::browseURL()}} to open the link.}

\item{...}{
  Arguments passed on to \code{\link[utils:browseURL]{utils::browseURL}}
  \describe{
    \item{\code{browser}}{a non-empty character string giving the name of the
    program to be used as the HTML browser.  It should be in the PATH,
    or a full path specified.  Alternatively, an \R function to be
    called to invoke the browser.

    Under Windows \code{NULL} is also allowed (and is the default), and
    implies that the file association mechanism will be used.
  }
    \item{\code{encodeIfNeeded}}{Should the URL be encoded by
    \code{\link[utils]{URLencode}} before passing to the browser?  This is not
    needed (and might be harmful) if the \code{browser} program/function
    itself does encoding, and can be harmful for \samp{file://} URLs on some
    systems and for \samp{http://} URLs passed to some CGI applications.
    Fortunately, most URLs do not need encoding.}
  }}

\item{package}{For \code{\link[=href_pkgdown]{href_pkgdown()}}; which package to link to.}
}
\value{
Always returns the URL, albeit invisibly if \code{open = TRUE}.
}
\description{
Returns URLs and, optionally, opens them in a browser.
}
